op3: Actually re-enable HAL3 capabilities
Due to the use of setprop on a persist* property, persist.camera.HAL3.enabled
remained persistent in userdata, so removing the setprop that caused this
isn't enough to get HAL3 back.

Forcefully enable HAL3 capabilities as early as possible into boot by
writing 1 to the persistent prop's path. We can't use setprop this early,
so manually writing the desired value is the only option.

Change-Id: Ie3c0a0727a4a010a9aabc5e64904f0b4f702cfb3